+++ FrannHammer.Api.Tests/Authentation/BaseAuthenticationTest.cs
@@ -11,6 +11,7 @@ using Autofac;
 using Autofac.Integration.WebApi;
 using Effort.DataLoaders;
 using FrannHammer.Api.Models;
+using FrannHammer.Api.Tests.Controllers;
 using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.EntityFramework;
 using Microsoft.Owin.Cors;
@@ -20,7 +21,7 @@ using Newtonsoft.Json.Serialization;
 using NUnit.Framework;
 using Owin;
 
-namespace FrannHammer.Api.Tests.Controllers
+namespace FrannHammer.Api.Tests.Authentation
 {
     public abstract class BaseAuthenticationTest : BaseControllerTest
     {
