+++ FrannHammer.Api.Tests/Authentation/CharactersAuthenticationTest.cs
@@ -2,20 +2,32 @@
 using System.Threading.Tasks;
 using NUnit.Framework;
 
-namespace FrannHammer.Api.Tests.Controllers.SmashAttributeTypes
+namespace FrannHammer.Api.Tests.Authentation
 {
     [TestFixture]
-    public class SmashAttributeTypeAuthControllerTest : BaseAuthenticationTest
+    public class CharactersAuthenticationTest : BaseAuthenticationTest
     {
-        private const string BaseUri = "/api/smashattributetypes";
+        private const string BaseUri = "/api/characters";
+
+        //[Test]
+        //[TestCase(BaseUri)]
+        //[TestCase(BaseUri + "/2")]
+        //[TestCase(BaseUri + "/4/movements")]
+        //[TestCase(BaseUri + "/10/moves")]
+        //public async Task ShouldGetUnauthorizedWithoutLogin_GET(string uri)
+        //{
+        //    var response = await GetAsync(uri);
+
+        //    Assert.That(response.StatusCode, Is.EqualTo(HttpStatusCode.Unauthorized));
+        //}
 
         [Test]
         [TestCase(BaseUri)]
         public async Task ShouldGetUnauthorizedWithoutLogin_POST(string uri)
         {
-            var smashAttributeType = TestObjects.SmashAttributeType();
+            var character = TestObjects.Character();
 
-            var response = await PostAsync(uri, smashAttributeType);
+            var response = await PostAsync(uri, character);
             Assert.That(response.StatusCode, Is.EqualTo(HttpStatusCode.Unauthorized));
         }
 
@@ -23,9 +35,9 @@ namespace FrannHammer.Api.Tests.Controllers.SmashAttributeTypes
         [TestCase(BaseUri + "/4")]
         public async Task ShouldGetUnauthorizedWithoutLogin_PUT(string uri)
         {
-            var smashAttributeType = TestObjects.SmashAttributeType();
+            var character = TestObjects.Character();
 
-            var response = await PutAsync(uri, smashAttributeType);
+            var response = await PutAsync(uri, character);
             Assert.That(response.StatusCode, Is.EqualTo(HttpStatusCode.Unauthorized));
         }
