+++ FrannHammer.Api.Tests/Controllers/CharacterControllerTest.cs
